game
It is a real simple zombie survival game. The gameplay is a mixture of puzzle and shooting. You’re the hero and equipped with a fireball gun which rebounds for a limited number of times. The zombies would stand still at their places and you have to shoot the ball (fireball) in such a way that it would bounce back and kill all the zombies successfully.